Correctly document supported SDK versions to Dart 3
The latest version of this package currently requires a Flutter version that also supports Dart 3 (e.g. Flutter 3.10+). Running pub get with lower versions of Flutter fails with package resolution conflict error that is not easy to understand. This change correctly documents the necessary SDK to successfully resolve this package, and fails with a meaningful error if it doesn't.